requirements.txt
fastapi
uvicorn
pydantic[email]
passlib[bcrypt]
python-jose
pyjwt
pymongo
pymongo[srv]
python-dotenv